add support for the AVIF, HEIC and HEIF formats
These 3 formats have in common that they use the HEIF container format specified in ISO/IEC 23008-12.

The MIME types and variant types (file extension) for HEIC and HEIF are based on sections C.2 and D.2 of the specification.

A HEIF container can contain multiple images, so the logic is kept as in `:ico` (take the size of the largest image).

The size of an image is found in the `ispe` box.